refactor: articleMapperをインフラに移動
