+++ src/domain/article/infrastructure/articleCommandServiceImpl.ts
@@ -2,7 +2,7 @@ import Article from '@/domain/article/model/article';
 import { ArticleCommandService } from '@/domain/article/repository/articleCommandService';
 import { ServerError, getErrorMessage } from '@/common/errors';
 import { AsyncResult, resultSuccess, resultError } from '@/common/types/utility';
-import fromPrismaToArticle from '@/domain/article/mapper/articleMapper';
+import fromPrismaToArticle from '@/domain/article/infrastructure/articleMapper';
 import { RdbClient } from '@/infrastructure/rdb';
 
 export default class ArticleCommandServiceImpl implements ArticleCommandService {
